# frontend/openapi.json
{
  "openapi": "3.1.0",
  "info": { "title": "OpenAPI definition", "version": "v0" },
  "servers": [{ "url": "http://localhost:18080", "description": "Generated server url" }],
...
  "paths": {
...
    "/api/v1/pricings/{id}": {
      "get": {
        "tags": ["Pricing"],
        "summary": "価格詳細を取得する",
        "operationId": "findPricingById",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "responses": {
          "200": {
            "description": "取得成功",
            "content": {
              "*/*": { "schema": { "$ref": "#/components/schemas/PricingDetailResponse" } }
            }
          },
          "404": {
            "description": "見つからない",
            "content": {
              "*/*": { "schema": { "$ref": "#/components/schemas/PricingDetailResponse" } }
            }
          }
        }
      },
      "put": {
        "tags": ["Pricing"],
        "summary": "価格を更新する",
        "operationId": "updatePricing",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/UpdatePricingRequest" }
            }
          },
          "required": true
        },
        "responses": { "200": { "description": "更新成功" } }
      },
      "delete": {
        "tags": ["Pricing"],
        "summary": "価格を削除する",
        "operationId": "deletePricing",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "requestBody": {
          "content": {
            "application/json": { "schema": { "$ref": "#/components/schemas/DeleteRequest" } }
          },
          "required": true
        },
        "responses": { "204": { "description": "削除成功" } }
      }
    },
...
    "/api/v1/pricings": {
      "get": {
        "tags": ["Pricing"],
        "summary": "価格一覧を取得する",
        "operationId": "listPricing",
        "parameters": [
          { "name": "productId", "in": "query", "required": false, "schema": { "type": "string" } },
          { "name": "level", "in": "query", "required": false, "schema": { "type": "string" } },
          { "name": "areaCode", "in": "query", "required": false, "schema": { "type": "string" } },
          {
            "name": "asOf",
            "in": "query",
            "required": false,
            "schema": { "type": "string", "format": "date-time" }
          },
          {
            "name": "page",
            "in": "query",
            "description": "Zero-based page index (0..N)",
            "required": false,
            "schema": { "type": "integer", "default": 0, "minimum": 0 }
          },
          {
            "name": "size",
            "in": "query",
            "description": "The size of the page to be returned",
            "required": false,
            "schema": { "type": "integer", "default": 20, "minimum": 1 }
          },
          {
            "name": "sort",
            "in": "query",
            "description": "Sorting criteria in the format: property,(asc|desc). Default sort order is ascending. Multiple sort criteria are supported.",
            "required": false,
            "schema": { "type": "array", "items": { "type": "string" } }
          }
        ],
        "responses": {
          "200": {
            "description": "取得成功",
            "content": {
              "*/*": { "schema": { "$ref": "#/components/schemas/PagePricingSummaryResponse" } }
            }
          }
        }
      },
      "post": {
        "tags": ["Pricing"],
        "summary": "価格を作成する",
        "operationId": "createPricing",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/CreatePricingRequest" }
            }
          },
          "required": true
        },
        "responses": { "201": { "description": "作成成功" } }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "UpdatePricingRequest": {
        "type": "object",
        "properties": {
          "amount": { "type": "number" },
          "validFrom": { "type": "string", "format": "date-time" },
          "validTo": { "type": "string", "format": "date-time" },
          "version": { "type": "integer", "format": "int32" }
        },
        "required": ["amount", "validFrom"]
      },
...
      "CreatePricingRequest": {
        "type": "object",
        "properties": {
          "productId": { "type": "string", "minLength": 1 },
          "level": { "type": "string", "minLength": 1 },
          "areaCode": { "type": "string", "minLength": 1 },
          "amount": { "type": "number" },
          "validFrom": { "type": "string", "format": "date-time" },
          "validTo": { "type": "string", "format": "date-time" }
        },
        "required": ["amount", "areaCode", "level", "productId", "validFrom"]
      },
...
      "PageableObject": {
        "type": "object",
        "properties": {
          "offset": { "type": "integer", "format": "int64" },
          "paged": { "type": "boolean" },
          "pageNumber": { "type": "integer", "format": "int32" },
          "pageSize": { "type": "integer", "format": "int32" },
          "sort": { "$ref": "#/components/schemas/SortObject" },
          "unpaged": { "type": "boolean" }
        }
      },
...
      "SortObject": {
        "type": "object",
        "properties": {
          "empty": { "type": "boolean" },
          "sorted": { "type": "boolean" },
          "unsorted": { "type": "boolean" }
        }
      },
...
      "PagePricingSummaryResponse": {
        "type": "object",
        "properties": {
          "totalElements": { "type": "integer", "format": "int64" },
          "totalPages": { "type": "integer", "format": "int32" },
          "size": { "type": "integer", "format": "int32" },
          "content": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/PricingSummaryResponse" }
          },
          "number": { "type": "integer", "format": "int32" },
          "pageable": { "$ref": "#/components/schemas/PageableObject" },
          "sort": { "$ref": "#/components/schemas/SortObject" },
          "first": { "type": "boolean" },
          "last": { "type": "boolean" },
          "numberOfElements": { "type": "integer", "format": "int32" },
          "empty": { "type": "boolean" }
        }
      },
      "PricingSummaryResponse": {
        "type": "object",
        "properties": {
          "id": { "type": "string" },
          "productId": { "type": "string" },
          "level": { "type": "string" },
          "areaCode": { "type": "string" },
          "amount": { "type": "number" }
        }
      },
      "PricingDetailResponse": {
        "type": "object",
        "properties": {
          "id": { "type": "string" },
          "productId": { "type": "string" },
          "level": { "type": "string" },
          "areaCode": { "type": "string" },
          "amount": { "type": "number" },
          "validFrom": { "type": "string", "format": "date-time" },
          "validTo": { "type": "string", "format": "date-time" },
          "version": { "type": "integer", "format": "int32" }
        }
      },
...
      "DeleteRequest": {
        "type": "object",
        "properties": { "version": { "type": "integer", "format": "int32" } }
      }
    }
  }
}